pyproject.toml
[tool.poetry]
name = "simple_app"
version = "0.1.0"
description = ""
authors = ["The Gang Technology"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.0"
flake8 = "^6.0.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
ipykernel = "^6.22.0"
flake8-annotations = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.poetry.scripts]
runserver = "simple_app.main:run"
